+++ include/envoy/buffer/buffer.h
@@ -238,7 +238,7 @@ public:
     auto result = static_cast<T>(0);
     constexpr const auto all_bits_enabled = static_cast<T>(~static_cast<T>(0));
 
-    char* bytes = reinterpret_cast<char*>(std::addressof(result));
+    int8_t* bytes = reinterpret_cast<int8_t*>(std::addressof(result));
     copyOut(start, Size, &bytes[displacement]);
 
     constexpr const auto most_significant_read_byte =