+++ source/exe/main_common.cc
@@ -59,7 +59,8 @@ MainCommonBase::MainCommonBase(OptionsImpl& options) : options_(options) {
     Thread::BasicLockable& log_lock = restarter_->logLock();
     Thread::BasicLockable& access_log_lock = restarter_->accessLogLock();
     auto local_address = Network::Utility::getLocalAddress(options_.localAddressIpVersion());
-    Logger::Registry::initialize(options_.logLevel(), options_.logFormat(), log_lock);
+    logging_context_ =
+        std::make_unique<Logger::Context>(options_.logLevel(), options_.logFormat(), log_lock);
 
     stats_store_ = std::make_unique<Stats::ThreadLocalStoreImpl>(options_.statsOptions(),
                                                                  restarter_->statsAllocator());
@@ -72,7 +73,8 @@ MainCommonBase::MainCommonBase(OptionsImpl& options) : options_(options) {
   }
   case Server::Mode::Validate:
     restarter_.reset(new Server::HotRestartNopImpl());
-    Logger::Registry::initialize(options_.logLevel(), options_.logFormat(), restarter_->logLock());
+    logging_context_ = std::make_unique<Logger::Context>(options_.logLevel(), options_.logFormat(),
+                                                         restarter_->logLock());
     break;
   }
 }