+++ source/extensions/filters/network/thrift_proxy/router/router_impl.cc
@@ -52,26 +52,8 @@ RouteConstSharedPtr RouteEntryImplBase::clusterEntry(uint64_t random_value) cons
   if (weighted_clusters_.empty()) {
     return shared_from_this();
   }
-
-  uint64_t selected_value = random_value % total_cluster_weight_;
-  uint64_t begin = 0UL;
-  uint64_t end = 0UL;
-
-  // Find the right cluster to route to based on the interval in which
-  // the selected value falls. The intervals are determined as
-  // [0, cluster1_weight), [cluster1_weight, cluster1_weight+cluster2_weight),..
-  for (const WeightedClusterEntrySharedPtr& cluster : weighted_clusters_) {
-    end = begin + cluster->clusterWeight();
-    ASSERT(end <= total_cluster_weight_);
-
-    if (selected_value >= begin && selected_value < end) {
-      return cluster;
-    }
-
-    begin = end;
-  }
-
-  NOT_REACHED_GCOVR_EXCL_LINE;
+  return WeightedClusterUtil::pickCluster(weighted_clusters_, total_cluster_weight_, random_value,
+                                          false);
 }
 
 bool RouteEntryImplBase::headersMatch(const Http::HeaderMap& headers) const {